chore(examples): regen bundled codegen for proptest_gen + adapter fixes
The prior six commits change codegen output for proptest harnesses (and
two adapter helpers used by the kani / guards backends), so the bundled
example artifacts drift against the new emission. Catches up the six
affected files so `qedgen check --regen-drift` stays clean.

Per-file rationale:

  examples/rust/lending/tests/proptest.rs        — adds `status:` line
                                                   to lifecycle init
                                                   seed-state (commit 3)

  examples/rust/multisig/tests/proptest.rs       — type-aware seed-state
  examples/rust/multisig/programs/tests/proptest.rs init for the
                                                   `voted: Map[32] U8`
                                                   field (commit 3) +
                                                   `status:` line

  examples/rust/percolator/tests/proptest.rs    — `accounts: Map[1024]
                                                   Account` gets the
                                                   record-aware default
                                                   `[Account::default();
                                                   1024]` (commit 3) +
                                                   `status: Status::
                                                   Active` (initial
                                                   lifecycle state, not
                                                   hardcoded
                                                   `Uninitialized`,
                                                   commit 3) + symmetric
                                                   i128 widening on
                                                   `pnl + s.I` add
                                                   guards (commit 1)

  examples/rust/percolator/tests/kani.rs        — symmetric i128
                                                   widening on the same
                                                   `pnl` add-guard
                                                   pattern (commit 1)

  examples/rust/percolator/programs/src/guards.rs — symmetric i128
                                                   widening on the
                                                   `(a + b) >= 0`
                                                   underwater check that
                                                   appears in `add_user`,
                                                   `add_lp`, `withdraw`,
                                                   and the three
                                                   `liquidate_*` arms
                                                   (commit 1)

(The escrow example is no longer affected — commit 4's gating means
mul_div helpers are only inlined when the spec uses them, which escrow
doesn't, and commit 6's gating means recursion_limit is only emitted for
specs with >32 fields, which escrow has 3.)

Drift detection: `qedgen check --regen-drift` is clean for these six
files; broader bundled-example drift unrelated to this branch is left
in place to keep the diff focused.

+++ examples/rust/percolator/programs/src/guards.rs
@@ -19,7 +19,7 @@ pub fn add_user<'info>(ctx: &mut AddUser<'info>, i: usize) -> Result<(), Program
     // requires: s.accounts[i].active = 0
     if !(ctx.vault.accounts[(i) as usize].active == 0) { return Err(ProgramError::from(PercolatorError::SlotAlreadyActive)); }
     // requires: (((s.accounts[i].capital) : Int)) + s.accounts[i].pnl ≥ (((0) : Int))
-    if !(((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ctx.vault.accounts[(i) as usize].pnl.get() >= ((0) as i128)) { return Err(ProgramError::from(PercolatorError::AccountUnderwater)); }
+    if !(((((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ((ctx.vault.accounts[(i) as usize].pnl.get()) as i128)) as i128) >= ((0) as i128)) { return Err(ProgramError::from(PercolatorError::AccountUnderwater)); }
     Ok(())
 }
 
@@ -31,7 +31,7 @@ pub fn add_lp<'info>(ctx: &mut AddLp<'info>, i: usize) -> Result<(), ProgramErro
     // requires: s.accounts[i].active = 0
     if !(ctx.vault.accounts[(i) as usize].active == 0) { return Err(ProgramError::from(PercolatorError::SlotAlreadyActive)); }
     // requires: (((s.accounts[i].capital) : Int)) + s.accounts[i].pnl ≥ (((0) : Int))
-    if !(((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ctx.vault.accounts[(i) as usize].pnl.get() >= ((0) as i128)) { return Err(ProgramError::from(PercolatorError::AccountUnderwater)); }
+    if !(((((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ((ctx.vault.accounts[(i) as usize].pnl.get()) as i128)) as i128) >= ((0) as i128)) { return Err(ProgramError::from(PercolatorError::AccountUnderwater)); }
     Ok(())
 }
 
@@ -85,7 +85,7 @@ pub fn withdraw<'info>(ctx: &mut Withdraw<'info>, i: usize, amount: u128) -> Res
     // requires: s.accounts[i].capital ≥ amount
     if !(ctx.vault.accounts[(i) as usize].capital.get() >= amount) { return Err(ProgramError::from(PercolatorError::InsufficientFunds)); }
     // requires: (((s.accounts[i].capital) : Int)) + s.accounts[i].pnl ≥ (((amount) : Int))
-    if !(((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ctx.vault.accounts[(i) as usize].pnl.get() >= ((amount) as i128)) { return Err(ProgramError::from(PercolatorError::InsufficientFunds)); }
+    if !(((((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ((ctx.vault.accounts[(i) as usize].pnl.get()) as i128)) as i128) >= ((amount) as i128)) { return Err(ProgramError::from(PercolatorError::InsufficientFunds)); }
     Ok(())
 }
 
@@ -149,7 +149,7 @@ pub fn liquidate_case_0<'info>(ctx: &mut LiquidateCase0<'info>, i: usize) -> Res
     // requires: s.accounts[i].active = 1
     if !(ctx.vault.accounts[(i) as usize].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
     // requires: (((s.accounts[i].capital) : Int)) + s.accounts[i].pnl ≥ (((0) : Int))
-    if !(((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ctx.vault.accounts[(i) as usize].pnl.get() >= ((0) as i128)) { return Err(ProgramError::Custom(0xFF)); }
+    if !(((((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ((ctx.vault.accounts[(i) as usize].pnl.get()) as i128)) as i128) >= ((0) as i128)) { return Err(ProgramError::Custom(0xFF)); }
     // requires: 0 = 1
     if !(false) { return Err(ProgramError::from(PercolatorError::AccountHealthy)); }
     Ok(())
@@ -163,9 +163,9 @@ pub fn liquidate_case_1<'info>(ctx: &mut LiquidateCase1<'info>, i: usize) -> Res
     // requires: s.accounts[i].active = 1
     if !(ctx.vault.accounts[(i) as usize].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
     // requires: ¬((((s.accounts[i].capital) : Int)) + s.accounts[i].pnl ≥ (((0) : Int)))
-    if !(!(((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ctx.vault.accounts[(i) as usize].pnl.get() >= ((0) as i128))) { return Err(ProgramError::Custom(0xFF)); }
+    if !(!(((((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ((ctx.vault.accounts[(i) as usize].pnl.get()) as i128)) as i128) >= ((0) as i128))) { return Err(ProgramError::Custom(0xFF)); }
     // requires: (((s.accounts[i].capital) : Int)) + s.accounts[i].pnl + (((s.I) : Int)) ≥ (((0) : Int))
-    if !(((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ctx.vault.accounts[(i) as usize].pnl.get() + ((ctx.vault.I.get()) as i128) >= ((0) as i128)) { return Err(ProgramError::Custom(0xFF)); }
+    if !(((((((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ((ctx.vault.accounts[(i) as usize].pnl.get()) as i128)) as i128) + ((ctx.vault.I.get()) as i128)) as i128) >= ((0) as i128)) { return Err(ProgramError::Custom(0xFF)); }
     Ok(())
 }
 
@@ -177,9 +177,9 @@ pub fn liquidate_otherwise<'info>(ctx: &mut LiquidateOtherwise<'info>, i: usize)
     // requires: s.accounts[i].active = 1
     if !(ctx.vault.accounts[(i) as usize].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
     // requires: ¬((((s.accounts[i].capital) : Int)) + s.accounts[i].pnl ≥ (((0) : Int)))
-    if !(!(((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ctx.vault.accounts[(i) as usize].pnl.get() >= ((0) as i128))) { return Err(ProgramError::Custom(0xFF)); }
+    if !(!(((((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ((ctx.vault.accounts[(i) as usize].pnl.get()) as i128)) as i128) >= ((0) as i128))) { return Err(ProgramError::Custom(0xFF)); }
     // requires: ¬((((s.accounts[i].capital) : Int)) + s.accounts[i].pnl + (((s.I) : Int)) ≥ (((0) : Int)))
-    if !(!(((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ctx.vault.accounts[(i) as usize].pnl.get() + ((ctx.vault.I.get()) as i128) >= ((0) as i128))) { return Err(ProgramError::Custom(0xFF)); }
+    if !(!(((((((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ((ctx.vault.accounts[(i) as usize].pnl.get()) as i128)) as i128) + ((ctx.vault.I.get()) as i128)) as i128) >= ((0) as i128))) { return Err(ProgramError::Custom(0xFF)); }
     // requires: 0 = 1
     if !(false) { return Err(ProgramError::from(PercolatorError::BankruptPosition)); }
     Ok(())
